5th_lab/requirements.txt
flask==2.3.0
gunicorn==21.0.0
prometheus-client==0.17.0
kafka-python==2.0.2
